workshops/crud_python/pymongo/requirements.txt
dnspython==2.5.0
pymongo==4.6.1
python-dotenv==1.0.1
